Processing image synchro between front and back; requirements bug correction; implement back workflow for dark; start working on dark workflow on frontend;

+++ back/app/models/processing.py
@@ -8,5 +8,5 @@ class ImageProcessing(BaseModel):
     b : float
     whites : int
     blacks : int
-    midtones : int
+    midtones : float
     stretchAlgo: int
